+++ Meraki.Api/Interfaces/General/Organizations/IOrganizationsApiRequests.cs
@@ -111,4 +111,88 @@ public interface IOrganizationsApiRequests
 		[AliasAs("adminIds[]")] List<string>? adminIds = null,
 		[AliasAs("userAgent[]")] List<string>? userAgent = null,
 		CancellationToken cancellationToken = default);
+
+	/// <summary>
+	/// Lists API request response codes and their counts aggregated by admin
+	/// </summary>
+	/// <remarks>
+	/// This endpoint is in beta.
+	/// GET /organizations/{organizationId}/apiRequests/responseCodes/history/byAdmin
+	/// </remarks>
+	/// <exception cref="ApiException">Thrown when fails to make API call</exception>
+	/// <param name="organizationId">The organization id</param>
+	/// <param name="t0">The beginning of the timespan for the data. The maximum lookback period is 31 days from today. (optional)</param>
+	/// <param name="t1">The end of the timespan for the data. t1 can be a maximum of 31 days after t0. (optional)</param>
+	/// <param name="timespan">The timespan for which the information will be fetched. If specifying timespan, do not specify parameters t0 and t1. The value must be in seconds and be less than or equal to 31 days. The default is 1 day. (optional)</param>
+	/// <param name="cancellationToken">The cancellation token</param>
+	[Get("/organizations/{organizationId}/apiRequests/responseCodes/history/byAdmin")]
+	Task<List<ApiResponseCodesHistoryByAdmin>> GetOrganizationApiRequestsResponseCodesHistoryByAdminAsync(
+		string organizationId,
+		string? t0 = null,
+		string? t1 = null,
+		double? timespan = null,
+		CancellationToken cancellationToken = default);
+
+	/// <summary>
+	/// Lists API request response codes and their counts aggregated by application
+	/// </summary>
+	/// <remarks>
+	/// This endpoint is in beta.
+	/// GET /organizations/{organizationId}/apiRequests/responseCodes/history/byApplication
+	/// </remarks>
+	/// <exception cref="ApiException">Thrown when fails to make API call</exception>
+	/// <param name="organizationId">The organization id</param>
+	/// <param name="t0">The beginning of the timespan for the data. The maximum lookback period is 31 days from today. (optional)</param>
+	/// <param name="t1">The end of the timespan for the data. t1 can be a maximum of 31 days after t0. (optional)</param>
+	/// <param name="timespan">The timespan for which the information will be fetched. If specifying timespan, do not specify parameters t0 and t1. The value must be in seconds and be less than or equal to 31 days. The default is 1 day. (optional)</param>
+	/// <param name="cancellationToken">The cancellation token</param>
+	[Get("/organizations/{organizationId}/apiRequests/responseCodes/history/byApplication")]
+	Task<List<ApiResponseCodesHistoryByApplication>> GetOrganizationApiRequestsResponseCodesHistoryByApplicationAsync(
+		string organizationId,
+		string? t0 = null,
+		string? t1 = null,
+		double? timespan = null,
+		CancellationToken cancellationToken = default);
+
+	/// <summary>
+	/// Aggregates API usage data by operationId
+	/// </summary>
+	/// <remarks>
+	/// This endpoint is in beta.
+	/// GET /organizations/{organizationId}/apiRequests/responseCodes/history/byOperation
+	/// </remarks>
+	/// <exception cref="ApiException">Thrown when fails to make API call</exception>
+	/// <param name="organizationId">The organization id</param>
+	/// <param name="t0">The beginning of the timespan for the data. The maximum lookback period is 31 days from today. (optional)</param>
+	/// <param name="t1">The end of the timespan for the data. t1 can be a maximum of 31 days after t0. (optional)</param>
+	/// <param name="timespan">The timespan for which the information will be fetched. If specifying timespan, do not specify parameters t0 and t1. The value must be in seconds and be less than or equal to 31 days. The default is 1 day. (optional)</param>
+	/// <param name="cancellationToken">The cancellation token</param>
+	[Get("/organizations/{organizationId}/apiRequests/responseCodes/history/byOperation")]
+	Task<List<ApiResponseCodesHistoryByOperation>> GetOrganizationApiRequestsResponseCodesHistoryByOperationAsync(
+		string organizationId,
+		string? t0 = null,
+		string? t1 = null,
+		double? timespan = null,
+		CancellationToken cancellationToken = default);
+
+	/// <summary>
+	/// Aggregates API usage by source IP
+	/// </summary>
+	/// <remarks>
+	/// This endpoint is in beta.
+	/// GET /organizations/{organizationId}/apiRequests/responseCodes/history/bySourceIp
+	/// </remarks>
+	/// <exception cref="ApiException">Thrown when fails to make API call</exception>
+	/// <param name="organizationId">The organization id</param>
+	/// <param name="t0">The beginning of the timespan for the data. The maximum lookback period is 31 days from today. (optional)</param>
+	/// <param name="t1">The end of the timespan for the data. t1 can be a maximum of 31 days after t0. (optional)</param>
+	/// <param name="timespan">The timespan for which the information will be fetched. If specifying timespan, do not specify parameters t0 and t1. The value must be in seconds and be less than or equal to 31 days. The default is 1 day. (optional)</param>
+	/// <param name="cancellationToken">The cancellation token</param>
+	[Get("/organizations/{organizationId}/apiRequests/responseCodes/history/bySourceIp")]
+	Task<List<ApiResponseCodesHistoryBySourceIp>> GetOrganizationApiRequestsResponseCodesHistoryBySourceIpAsync(
+		string organizationId,
+		string? t0 = null,
+		string? t1 = null,
+		double? timespan = null,
+		CancellationToken cancellationToken = default);
 }